Add some basic code for generating unit ids.
This little module contains two methods for creating and decoding unit ids.
This does not generate them from CIF files directly, instead this will generate
them from dictionaries which can be computed from CIF data. In addition, this
can handle generating a short id, that is one with the optional fields left
out. While this generates atom ids it does not have the logic to check that the
all fields required for an atom id are present.