Incomplete implementation of Components
A component can filter and order the atoms it is made of. This allows for the
filtering but not yet ordering because the implementation is incomplete. The
implementation is generic since we want to do the same thing with a structure and
the residues or units it is made of. The filtering allows for filtering by an
exact string, a list or a generic function.