Outline CIF reader
This is the generic outline for how the CIF reader will work. Not all methods
are implemented yet. One problem we may run into is that storing all symmetries
for a large structure may be very expensive. In the future we may have to only
read the asymetric unit and then generate the transformed positions as needed.