Implement ordering in EntityContainers
We can now get entities from and EntityContainer in a specified order. We can
also filter and order in one call. This will make it easy to do things like get
all heavy chain atoms in a given order from a residue. The implementation could
probably be more efficient but this is good enough for now.