Code clean up
I've gone through and organized the code to be a bit more pythonic. This
includes small to large changes. Small changes are things like having spaces
after commas in function calls, changing lines to be less than 80 characters,
and only importing at the top of the package. Some of the larger changes are
like reorganizing the superposition tests so we follow more standard python
conventions. I also deleted the duplicate superpositions.py file in tests/ as
it should be part of FR3D and not the tests. Lots of white space has been
removed and whitespace around operators and commas has been added.

I also moved the documentation for the function into a doc string for the
function. This is the standard way to document functions in python. Someone
should check that the documentation is correct.

Along the way I looked at the RMSD calculation and I'm not sure what it is
doing. It is computed but never used. I can comment it out and all tests pass
however I'm leaving it in with a note for now until we understand why it is
there. This also showed me that E0 is only used by RMSD so it too could be
removed if we don't need RMSD.

In addition, the code to test if the reflection is -1 was odd. When
comparing floats the thing to do is to get the difference and see if it is
small enough. numpy has a function for this, isclose, so we use that instead.