Add a basic is_complete method to Entity
This checks if an entity is complete by looking to see if we find the expected
number of atoms with the given name. It is possible to use a value other than
name however. The key used must provide a unique value for each atom though. If
not then it is possible that the method will not work correctly. For example,
if the key provides two things in one case and one is missing then the total
count will be correct and thus the returned value will be True when it should
be False.