Changed the first two tests, and last test
For the first two tests, the answer matrix, 'ans', was previously the
rotation matrix from b to a.  The function 'besttransformation' provides
the rotation matrix from a to b.  Thus, the inverse of 'ans' is the
correct matrix to be comparing to.  Recall, the inverse of a rotation
matrix is it's transpose.

For the last part, there was an error in the test because of the way I
asked for the trigonometric functions to be evaluated.