Setup to compute residue centers
We now have a place in components to store centers. That is the .centers
dictonary. I have also added a method called __compute_centers__ which will
compute the center coordinates using the mean of the requested atoms. This
method should be called during construction in the __init__ method to compute
the requested centers. Right now it only computes the base center using heavy
atoms but this can be extended.