Use new mmCIF reader
We are now using PyCIFRW instead of PDB's cif reader. This is useful because
this package is distributed widely making it easier to give our code to other
people. The overall api to the fr3d CIF reader is the same because the only
changes have been internal to that module.