Use keyword arguments to construct and reorganize
Atom and Component are now built using keyword arguments. The test code and
infer hydrogen code from the previous commit did not do this. This fixes that.
In addition, we move infering the hydrogens into its own test class. This is
because we will later do more detailed tests to check that the hydrogens are in
the right position so we want to have a common setup for them. In addition, we
fix a few minor problems with mixed tabs and spaces as well as some whitespace
issues.