Add a AtomProxy object
The AtomProxy serves as a dictionary like container to get and store center
coordinates. However, it will use the coordinates of atoms if given the name of
an atom and there is no value for that key. This allows for simpler code in
places, like discrepancy, where we will want either positions of centers of
various types as well as atoms.

This commit adds an implementation of that object as well as some tests for it.
It's a bit short on documentation now though.