Build AtomProxy with the component
This was done because if we add atoms to the component this is not reflected in
the AtomProxy. Instead of trying to update both, I just have the AtomProxy
refer to the residue it belongs to and use it's ._atoms list of atoms. This
means that when we infer_hyrdogens we then go an use an infered hyrodgen as a
center as well.