Check that proxy works when _atoms is modified
The proxy should work if the _atoms attribute in it's component is modified.
That attribute will be modified when doing infer_hyrdogens but may be done at
other times, for example we may try to infer some missing atoms and would still
like to be able to treat them as centers.