Added if statemnt to discrepancy.  This gets rid
of the doubling.  Also, when you use discrepancy function, you must
specify what weights you are talking about.  An input can't be
[1,1,1,0,0].  You must specify, P_weights = [1,1,1,0,0].  Otherwise,
P_weights is always assumed to be all ones.  After this simple fix, the
rotation matrices now match for the last two examples, and the
discrepancy that is to be near 0 is in fact near 0.  However, the
discrepancy still does not match for the last two examples.