Generate the high level structure
This can now build the high level data structures from cif fairly correctly. It
does not yet apply the required transformations for symmetry operators however.
In addition, I have changed the residues in chain to be an iterator as there are
non polymeric entries attached to chains such as water and ions. While we
maintain the attachment to a chain we do not iterate over them by default. This
is done by setting polymeric to False by default in the iterator. However if we
set it to None we can iterate over these entries as well. In addition, I've
corrected some ordering issues when building the list of chains.