Attempt at computing simple transforms
This should correctly compute the transformed coordinates for simple
transformations. This can apply the transformation matrix but does not use the
transformation vector as I don't quite get how to use that yet. However, this
seems to find transformed coordinates correctly. I have added some simple tests
to the high level interface to examine this.