Compute experimental sequence mapping
We can now generate experimental sequence mapping for chains in a cif file.
Currently we only use chains from symmetry 1_555 and model 1. This is a bit
limited but still useful. I've also fixed a small problem in the unit id
generation logic. That module will be updated to use the newer one from rna
structure later.