Modify atom proxy to accept a tuple of atoms
This will allow people to access the coordinates for a set of atoms by name, not
just an individual atom or a predefined one. However, this will fail silently if
given a name for an unknown atom. This is not good but should be fixed. In
addition, this makes the current len implementation limited. I'm leaving the len
alone for now as it is reasonable and having this iterate over all possible
combinations seems foolish. We also cannot do the center of a predefined thing
as well as that of atoms at once.