Add a cifatom writer and a simple smoke test
I've added an initial cif atom writer class and trivial test for it. A cifatom
file is basically only the atom_site entry in a cif file with some minor
differences. The most notable are we include all atoms in the cif file, even
those produced by rotations and we append unit ids to each entry. There are more
differences but these are detailed in the documentation for the cifatom class.
In addition, I added a crude test to make sure it runs and produces possible
output. I need a more through check of this functionality though.