Add tests for building the transformation matrix
When computing the symmetry operated coordinates it is best to have a
transformation matrix as described in:

http://www.ncbi.nlm.nih.gov/pmc/articles/PMC2677383/

This is basically a 4x4 matrix where the first 3x3 grid is the symmetry matrix
and the third column is the symmetry vector and 1.0 in the 3,3 position. The
rest of the last row is 0. I've added some tests to the low level interface for
a structure (4NGG) that has an interesting transformation matrix. We will use
this matrix to fix the issue where we do compute the correct coordinates for
this structure.