Fix computation of symmetry operated coordinates
I was computing these operations incorrectly because I did not use the
transformation vector. I'm now using it and am hopefully making the correct
transformations. I'm using the values provided by Dr. Zirbel as tests for 4NGG
though this makes it seem like my values for 1WMQ were wrong. This is possible
and likely so I've changed the expected values.