Add some ability to persist parsed cif data
We want to be persist parsed cif data. I have added a bit of code which should
do that as well as tests to go with it. Right now I do this by using pickle to
dump the data produced by PdbxReader. This requires that I mess with the
internal structure because it retains a link to the filehandle the file was
built from. I remove that to persist the data. This is not ideal and should be
fixed. I believe pdbx has it's own persist module so I should look into using
that.

The tests I added are a copy of the low level cif tests. I would like to have
this done through inheritance instead of copy paste but I don't get how nose
runs test well enough. I will look into that later.