Correct setting the default operators and naming
We need to set a default operator if we are dealing with a complex operator.
I was doing it incorrectly, where I only returned the id of the identity
operator. In addition, If we are dealing with one of the point groups, there is
no name available. We deal with this by setting the name to P(n), where n is the
point group number.