Fix bug with multiple symmetry operators
There was an issue where we would not always generate the correct number of
residues because we were splitting up the atoms that belonged in one residue
over multiple residues. This occurred because the atoms were not returned
sorted. This happened in some structures where we had more than one symmetry
operator for atoms. I've dealt with this by flipping around the logic of
building atoms. Basically we go through all atoms and then apply the first
operator, then the second and so forth. This is done by using itertools to
filter and map and so that we do not have to build large arrays. This does
change the ordering but overall everything works the same.