Fix problems with multiple symmetry operators
There are cases where things end up having the same symmetry operator applied
more than once. This is not good as we end up producing duplicate units. This is
not good as we should only ever have one unit in a structure. We get around this
by only returning the unique operators. It may be better to only store the
unique ones, but that can be added later. In addition, there is an issue where
we apply the default identity operator too many times. This is also dealt with,
by not adding it unless there are no other operators.