Change how experimental sequence mapping works
It should not use auth_seq_num as that may not correspond to the number in
coordinates. Instead we use the pdb number as that seems to work better. I found
an example 1I9K where the author numbers are always 0, but the pdb provided ones
seem to be correct.