Move tests around and fix another mapping bug
I've moved the low level tests where they belong. In addition, I found a bug
where since we assumed the symmetry operator for the unit in question is 1_555
when doing the experimental mapping we did not generate the correct mappings.
This is fixed by picking the first possible symmetry operator for the unit.
However, we should either generate mappings for all of them, or map to a unit id
that corresponds to all units related by symmetry/models.