Speed up experimental sequence mapping
There were several issues that made that method slow. Probably the worst was
that we fetched the model to use each time. That's unneeded and expensive, we
can just do it once. In addition, I've added a use of itertools and made it an
iterable. I may make it more of an iterable as time goes on, but right now this
is reasonably fast.