Add a transform method
We can now transform a component or atom to produce a new component or atom.
This is useful when we are comparing the positions of atoms and components for
clustering interactions and the like. I've also added some tests for this code.
At the moment this does preserve the unit id, which may be dangerous as it
becomes possible to confuse the rotated/moved coordinates with the original one.