Add a lookup method to atomproxy
This is a way to get the center for atoms, but without failing if atoms are
missing. This is important in some structures where not all atoms are present.
This is similar to get in that it tolerates missing atoms but it will return a
center for what it can compute if possible. If nothing is present then an empty
list is returned.