Add more tests to define the behavior
It should now be clear that accessing once stores the values even in a case
where there are missing atoms. This may lead to odd behavior where a lookup will
let a future access which should fail pass. This is the current behavior and may
be labeled as a bug in the future.