Large refactoring to a newer style
This style should be a bit easier to understand. This also adds useful things
like a pairwise iterator, unit ids for models or chains, and the ability to deal
with alternative ids when reading cif files. This fixes several bugs as well as
speeds things up a little. We no longer need to generate model and chain objects
as they aren't too useful and can be generated on demand using a selection of
residues.