First attempt classifiers
This adds a generic classifier and an empty one for RNA protein interactions.
This is something that @poorna11 should implement. The idea is to provide a
single method which takes as input the first and second component and produces
as output the classifications for that pair. This is untested but roughly provides
a framework to work in.