Store definitions of centers
It is useful in several cases to know how a center was defined. This allows that
to happen by adding a _definitions property to centers. We also add a method to
recover the definition. Also fix a small bug where * would not be recognized a
known name when it is hardcoded as being known.