Revert to using P_*
We should use P_* for the names of symmetry operators of point symmetry groups
in viruses. This is from a discussion with Dr. Zirbel and Neocles today. This is
consistent with our old naming scheme as well as looking like the names of
crystal symmetry operators. This indicates the relationship to crystal symmetry
operators.