Add test for observed but in component unit id
I found while running the pipeline that some unit ids would have a '.' in them
for the alt id. The cause of this was not setting the alt id to None for atoms
with a '.' alt id. This is already under tested and fixed. Now I add a test for
the observed issue to help prevent it in the future.