Allow experimental sequence mapping to use alt ids
This now lets an experimental sequence mapping use alt ids. The change is that
we have to build up a mapping of a key that we can compute from
pdbx_poly_seq_scheme to unit ids. This forces us to have a potentially very
large dictionary but there is no way around it now. As a result though we do get
the correct model and other things without any odd logic. This changes several
tests as they now produce many more mappings due to alt ids.