Improve consistency of returned data
The AtomProxy would sometimes return an empty list instead of an empty
numpy array. This proves troublesome when we have to check if the array
is true or not as numpy defines truth in arrays differently. TO avoid
this we always return a numpy array but when things are missing it gives
an empty list.