Begin some work to add coordinate trees
These are simple wrappers around KDTree's in scipy, which allow for
rapid querying and accessing by distances. This is a fairly common step
in annotating pairs so it makes sense to have an efficient way of doing
this. I've outlined a rough wrapper but have not tested it out yet. The
idea is to provide a slightly nicer interface where we can return atoms,
or residues instead of just indexes in the given array which is what the
standard tree's do.

However, this is also a large change. In doing we *must* drop support
for 2.6, as scipy does not support 2.6. While this shouldn't be a huge
burden since 2.6 is not supported by the python devs anymore either we
still use it on our servers. Thus we should also upgrade the servers
before this can be finished. For now this will stay on a separate
branch until we decide that we can move to 2.7.