Extend experimental_sequence_mapping to take lists
This method is used in the pipeline to compute experimental sequence
mappings. However, it is causing a very large slow down in the pipeline.
I think this is because each time we call it we have to recompute all
the residues each time as well as reparse the entire
pdbx_poly_seq_scheme data. This is not good.

We could deal with it by storing residues and the like, but I generally
try to avoid that as we can end up with a lot of data stored and eat a
lot of memory. Plus that wouldn't speed up the time a lot, I think.
Instead I extend the method to take a list of chains to process. This
should allow us to do all chains at once and reduce the overall time
without increasing memory too much.