Correct and matrix discrepancy function
We can now compute discrepancies given only the matrices required. I've
also added some tests to make sure it works the same as the current
discrepancy function. The interface isn't the saay yet but is close
enough for now, I think. Notably, the weight parameter is different and
this should probably be corrected.