Correct logic for setting index of components
The logic was flawed where if the seq_id is '.' then it would not set an index.
In this case we set it to None as that is what '.' means. We also test this
this only happens for non polymeric entries as that is the only time it should
happend.