Fix issue where some units are ignored
Some units were ignored if they did not have a symmetry operator
assigned. This happens in 2UUA when the unit is part of a asym_id that
has not been assigned to a assembly_gen. This occurs for chain Z
(contains ions and paromomycin).

We fix this by using all symmetry operators if we can't find one. This
may create too many operators but it will at least make sure the unit is
not ignored. We also log this state so hopefully I can see how common
this is.

Finally, I've emailed John Westbrook about this. The logic of what to do
may change depending upon his feedback.