Attempt to fix standard_transformation
This method was having some issues not computing the correct matrices.
I've tried to rewrite it so that the matrix passes some simple tests.
I have some concerns that I've had to lower the decimal value for the
almost equal. This may be because the matrix has only a few decimal
places and so errors build up. At least this appears to be usable now
however.