Improve transformation matrix computation
The base centers were slightly wrong because they counted all atoms that
are part of a standard base not just the heavy atoms. We define the
center of a base only considering the heavy atoms, however the previous
version included the non heavy atoms (like hydrogens) this lead to
errors in the computation of the translation vector. This is now fixed.
The rotation matrix still has errors in it and I don't yet know how to
fix them.