Correct alt id reading
This reworks how alt ids are handled. The correct handling is that if
there are alt ids and there are atoms without an assigned alt id it is
common to all components. Thus if there is say:

alt_id atom
None   C1
A      C2
B      C2

Then the C1 atom belongs should go to both alt id A and B. Also, there
is then no component with alt id None.

I reworked the reading to do this correctly. The idea is to no longer
group by unit id (which includes alt id) but by all parts of a unit id
but the alt id. Then we group the atoms by alt id. If there are atoms
with a None alt id we copy them into all other units with updated alt
ids.