Handle edge case of single atom residues The pipeline stage units.coordinates was terminating abnormally when addressing residues containing only one atom.  The data was being presented as a list, instead of a list of lists containing only a single list.
This change adds an argument to the CifAtom class,
protect_lists_of_lists, to add a dummy list and ensure that any
output that exits the routine contains at least two lists in its
list of lists.  The dummy list uses the existing line-skipping
logic in units.coordinates to prevent the dummy list from being
parsed further at that stage.

The argument is configured with the default of False to preserve
the pre-existing behavior for function calls which were behaving
normally with respect to single atom residues (primarily
export.cifatoms).