Put back old code for processing of old structures
We are trying to add glycosidic bonds to our database. There are around 25 structures that need to have their old naming conventions in order to be added to our database. This commit reinserts the code that existed before complex symmetries were handled so that we can get the same unit_ids. Otherwise, the code will default to deal with complex symmetries in a new way.