config and command outline
Initial addition of recommended file structure and addition of config.py file. 
Explanation of python scripts that can be ran from the command line and their required and optional flags. 
Changed some instances of naming but there are more instances of names that should be changed to more concise things such as .cif to 3d structure file since we're expanding to use pdb files too.